+++ weex_core/Source/android/jniprebuild/jni_load.cc
@@ -3,7 +3,7 @@
 #include <android/base/jni/android_jni.h>
 #include <android/bridge/impl/measure_mode_impl_android.h>
 #include <android/bridge/impl/content_box_measurement_impl_android.h>
-#include <android/bridge/impl/WXJsFunction_Impl_Android.h>
+#include <android/bridge/impl/jsfunction_impl_android.h>
 
 jint JNI_OnLoad (JavaVM *vm, void *reserved)
 {
